Analysis using Navajo Water Export data
Water quality is a major issue on Native American reservations in the southwestern United States. The
prevalence of uranium mines and uranium mill accidents mean that much of the water in the Navajo Nation
is irradiated, and many homes are left without clean, drinkable water. Multiple environmental agencies
routinely sample water in the region and report on contaminants.