DBLP Data using SQL connection
DBLP database of bibliographic information on major computer science
journals and proceedings, available from https://data.mendeley.com/datasets/3p9w84t5mr. The dataset has
been processed to include predictions of the author’s genders using the open-source Genderize API. The
processed data has been made available in the form of SQL scripts that import the data into a MySQL
database. We are primarily interested in the “general” and “authors” tables created by the “main.sql” and
“authors.sql” scripts, respectively.

Problem 1
Filter the data to include only the authors for whom a gender was predicted as ‘male’ or ‘female’ with a
probability of 0.90 or greater, and then create a bar plot showing the total number of distinct male and
female authors published each year. Comment on the visualization.

Problem 2
Still including only the authors for whom a gender was predicted with a probability of 0.90 or greater, create
a stacked bar plot showing the proportions of distinct male authors vs. distinct female authors published each
year. (The stacked bars for each year will sum to one.) Comment on the visualization.

Problem 3
Still including only the authors for whom a gender was predicted with a probability of 0.90 or greater, create
a faceted bar plots showing the proportions of female first authorships for each year for each domain (CS, DE,
SE, and TH). (If a conference belongs to multiple domains, you may include it for both of them.) Comment
on the visualization. Which domains have the highest and lowest representation of papers with female first
authors?