Prediction of Heart disease
The objective of this project is to build a predictive model to detect the presence of heart disease in a patient using the Cleveland Hospital data. The focus will be distinguishing the presence value (1) from the absence value (0) of a heart disease. Classification using Logistic regression and decision trees is implemented.